feat: Title claims, _resolve_bulk, hoist educational_text to Title.description
- Add claims (GenericRelation) to Title model for provenance-tracked fields
- Add _resolve_bulk() for efficient bulk claim resolution (single query +
  in-memory resolution + bulk_update) with FK handler support
- Hoist educational_text from MachineModel to Title.description via claims;
  sign copy MainText targets Title, Sources/Notes goes to claim citation
- Convert all ingestion commands to use _resolve_bulk instead of direct writes
- Refactor resolve.py into resolve/ package (_entities, _helpers, _relationships)
- Rename generate_ipdb_titles → ingest_ipdb_titles for naming consistency
- Show title_description on model page only for single-model titles
- Remove educational_text/sources_notes fields from MachineModel
