+++ backend/apps/catalog/management/commands/ingest_all.py
@@ -2,7 +2,7 @@
 
 Runs: ingest_pinbase_taxonomy → ingest_pinbase_manufacturers →
       ingest_pinbase_corporate_entities → ingest_pinbase_systems → ingest_ipdb →
-      ingest_opdb → generate_ipdb_titles → ingest_pinbase_models →
+      ingest_opdb → ingest_ipdb_titles → ingest_pinbase_models →
       ingest_pinbase_series → ingest_pinbase_titles → ingest_pinbase_signs →
       resolve_claims.
 """
@@ -20,7 +20,7 @@ STEPS = [
     "ingest_pinbase_systems",
     "ingest_ipdb",
     "ingest_opdb",
-    "generate_ipdb_titles",
+    "ingest_ipdb_titles",
     "ingest_pinbase_models",
     "ingest_pinbase_series",
     "ingest_pinbase_titles",
